+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Implemented FEP-5219 `affiliations` collection.
 - Added `configuration.accounts.max_note_length` field to `Instance` entity.
 - Added `configuration.accounts.max_display_name_length` field to `Instance` entity.
+- Maintain record of the total number of voters in a poll.
 
 ### Changed